Fix non-emulated computations at RPE_BITS >= 23.
If RPE_BITS was set to 23 or more then no truncation at all would
happen and full double precision would be retained. This is not
desirable as one may want to set RPE_BITS=23 to emulate single
precision. This change ensures that a value is always reduced to
at least single precision when reduce_precision is called.